avoiding create profiles during dry import (cuz of signal call)

+++ apps/aklub/models.py
@@ -1042,19 +1042,22 @@ def Userprofile_administrative_unit_changed(sender, **kwargs):
     user = kwargs['instance']
     units = kwargs['pk_set']
     action = kwargs['action']
-    if action == 'post_add':
-        # if user has 0 administrative units we set him as active because he was inactive before
-        choose_if_active(True)
-        for unit in units:
-            Preference.objects.get_or_create(
-                user=user,
-                administrative_unit_id=unit,
-            )
+    # dry_run is defined during importing..
+    # we want to call signal only if dry_run => false or doesnt exist
+    if not hasattr(user, 'dry_run') or not user.dry_run:
+        if action == 'post_add':
+            # if user has 0 administrative units we set him as active because he was inactive before
+            choose_if_active(True)
+            for unit in units:
+                Preference.objects.get_or_create(
+                    user=user,
+                    administrative_unit_id=unit,
+                )
 
-    elif action == 'post_remove':
-        user.preference_set.filter(administrative_unit__id__in=units).delete()
-        # if user has 0 administrative units we set him as inactive
-        choose_if_active(False)
+        elif action == 'post_remove':
+            user.preference_set.filter(administrative_unit__id__in=units).delete()
+            # if user has 0 administrative units we set him as inactive
+            choose_if_active(False)
 
 
 class Preference(models.Model):
